Add files via upload
This project is a tribute page dedicated to Imran Khan, showcasing his influential career and contributions to cricket and politics. Designed with HTML, CSS, the page features a responsive layout and engaging design elements that highlight key milestones, achievements, and philanthropic efforts. The project emphasizes front-end development skills and the creation of a visually appealing and informative web experience.