Add files via upload
This project features a sleek and modern landing page designed with HTML, CSS. The page is crafted to provide an engaging and user-friendly experience, with a focus on clean design, intuitive navigation, and responsive layout. Key elements include a visually appealing hero section, interactive components, and smooth scrolling effects. This project showcases skills in front-end development and design, demonstrating the ability to create effective and visually striking landing pages.